refactor: ♻️ remove code duplication for footer
