+++ app/src/components/students/list/client.tsx
@@ -4,11 +4,11 @@ import { cn } from '@pedaki/design/utils';
 import type { Field } from '@pedaki/services/students/query.model';
 import { ColumnSelector } from '~/components/datatable/column-selector';
 import { DataTable } from '~/components/datatable/data-table';
+import Footer from '~/components/datatable/footer';
 import type { StudentColumnDef, StudentData } from '~/components/students/list/columns.tsx';
 import { generateColumns } from '~/components/students/list/columns.tsx';
 import Filters from '~/components/students/list/filters.tsx';
-import Footer from '~/components/students/list/footer.tsx';
-import { searchParams } from '~/components/students/list/parameters.ts';
+import { searchParams, serialize } from '~/components/students/list/parameters.ts';
 import { useScopedI18n } from '~/locales/client.ts';
 import { api } from '~/server/clients/client.ts';
 import { useStudentsListStore } from '~/store/students/list/list.store.ts';
@@ -140,6 +140,8 @@ const Client = ({ className }: React.HTMLAttributes<HTMLDivElement>) => {
         meta={meta}
         sorting={sorting}
         columnVisibility={columnVisibility}
+        perPageLabel={t('footer.perPage')}
+        serialize={serialize}
       />
     </div>
   );
